### 2025-11-30 20eb209
Fixed a bad reference to current_lob in before_lob_update
