### 2025-11-30 f994583
chore: bump version to 0.2.8
