### 2026-01-09 2f03250
Fix AttributeError in OrderExecutedPriceMessage: use execution_price instead of price
The ITCH 5.0 OrderExecutedPriceMessage class uses `execution_price` as its
attribute name, but the code was incorrectly referencing `message.price`.
This fixes the AttributeError when processing OrderExecutedPriceMessage.

Fixes #69

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
