### 2026-01-09 50a61cf
chore: bump version to 0.2.9
