### 2026-01-09 016b3f5
chore: bump version to 0.2.10
