### 2026-01-20 d095814
Edited the paper to include the newly required sections.
