+++ paper/paper.md
@@ -33,9 +33,15 @@ MeatPy is a Python framework specifically developed for processing and analyzing
 
 High-frequency financial markets now operate at sub-millisecond time scales, with individual order placements, cancellations, and executions occurring in nanoseconds. To understand market microstructure, liquidity provision, and price formation, researchers increasingly rely on historical order book data that record every message sent by the exchange. Several exchanges, including Nasdaq, IEX, and the Australian Stock Exchange and Chi-X Australia, make their high-frequency data feeds available for academic research, often at no cost or a reduced fee.^[Nasdaq data can be obtained through their ["Academic Waiver Policy"](https://www.nasdaqtrader.com/content/AdministrationSupport/Policy/ACADEMICWAIVERPOLICY.pdf); IEX provides free historical data, called HIST, on a T+1 basis on their [website](https://iextrading.com/trading/market-data/#hist-download); ASX and Chi-X Australia data can be accessed via [SIRCA](https://sirca.org.au/) by their academic subscribers] These feeds offer unprecedented granularity, capturing not only trades and their direction, but also order placements, cancellations, halts, circuit breakers, and specialized exchange-specific order types. As such, they have been used to study a wide range of market microstructure-related questions [see, e.g., @oharaWhatsNotThere2014;@comerton2019inverted;@shkilkoEveryCloudHas2020;@gregoire2022earnings.]
 
-Despite their common conceptual structure, exchange data feeds differ in message formats, order type definitions, and exchanges rules which affect processing logic. A single day of Nasdaq TotalView ITCH data can exceed ten gigabytes in compressed binary form and contain billions of messages. Processing these heterogeneous and high-volume data streams has traditionally relied on writing custom scripts and programs which are kept private by the research team [see, e.g. @clarkmcgeheeObjectOrientedLibraryRealTime2013;@gaiFastConstructionNanosecond2013] or on using third-party processing tools or platforms such as LOBSTER by @huangLOBSTERLimitOrder2011.^[[LOBSTER](https://lobsterdata.com/) is available to academic institiutions as a flat-fee online service.] This situation limits reproducibility and slows research progress.
+Despite their common conceptual structure, exchange data feeds differ in message formats, order type definitions, and exchange rules which affect processing logic. A single day of Nasdaq TotalView ITCH data can exceed ten gigabytes in compressed binary form and contain billions of messages. Processing these heterogeneous and high-volume data streams has traditionally relied on custom scripts kept private by research teams or on proprietary third-party platforms. This situation limits reproducibility and slows research progress.
 
 MeatPy addresses these challenges by providing an open-source Python framework for parsing and analyzing high-frequency financial market data. It reconstructs full limit order books from raw feed data, supports multiple exchange-specific message formats,^[MeatPy currently supports TotalView-ITCH versions 4.1 and 5.0 used by exchanges on the INET platform which includes most Nasdaq-operated equity exchanges. Support for IEX DEEP+ is under developement.] and leverages modern Python features such a generators, context managers, and type annotations to improve reliability and developer productivity. By abstracting away the complexity of heterogeneous feed formats and offering efficient data processing primitives, MeatPy enables researchers to focus on designing and executing market microstructure analyses rather than building low-level data engineering infrastructure.
+
+# State of the Field
+
+Several tools exist for processing limit order book data. LOBSTER [@huangLOBSTERLimitOrder2011] is a widely-used commercial platform that reconstructs order books from Nasdaq ITCH data as an online service, but its closed-source nature limits extensibility and reproducibility. Other published implementations [@clarkmcgeheeObjectOrientedLibraryRealTime2013; @gaiFastConstructionNanosecond2013] were not released as reusable open-source libraries.
+
[…]
