### 2026-01-20 6e9eb77
Updated research impact
