### 2026-01-20 #8882 @vgreg
We updated the paper according to the new requirements.
