### 2026-01-20 a7df84b
chore: bump version to 0.2.11
