### 2026-01-24 8fa3315
feat: add support for older ITCH versions (2.0, 3.0, 4.0)
Implements GitHub issue #80 by adding three new ITCH protocol modules:

- ITCH 2.0: ASCII format with 6 message types, embedded timestamps (ms)
- ITCH 3.0: ASCII format with 16 message types, separate timestamp messages
- ITCH 4.0: Binary format with 6-character stock symbols (vs 8 in v4.1)

Each module provides MessageReader, MarketProcessor, and MarketMessage
classes following the same patterns as existing ITCH 4.1 and 5.0 modules.

Closes #80

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
