+++ src/meatpy/itch4/itch4_market_message.py
@@ -1,0 +1,1278 @@
+"""ITCH 4.0 market message types and parsing.
+
+This module provides classes for parsing and representing ITCH 4.0 market data
+messages. ITCH 4.0 is a binary format similar to ITCH 4.1 but uses 6-character
+stock symbols instead of 8-character symbols.
+
+Message Types:
+- T: Seconds
+- S: System Event
+- R: Stock Directory
+- H: Stock Trading Action
+- L: Market Participant Position
+- A: Add Order (no MPID)
+- F: Add Order (with MPID)
+- E: Order Executed
+- X: Order Cancel
+- D: Order Delete
+- U: Order Replace
+- P: Trade
+"""
+
+import json
+import struct
+
+from ..message_reader import MarketMessage
+
+
+class ITCH4MarketMessage(MarketMessage):
+    """A market message in ITCH 4.0 format.
+
+    This is the base class for all ITCH 4.0 message types. ITCH 4.0 uses
+    binary format with 6-character stock symbols.
+
+    Attributes:
+        timestamp: The nanosecond timestamp offset within the current second
+        type: The message type identifier
+        description: Human-readable description of the message type
+        message_size: The size of the message in bytes
+    """
+
+    type: bytes = b"?"
+    description: str = "Unknown Message"
+    message_size: int = 0
+
+    sysEventCodes = {
+        b"O": "Start of Messages",
+        b"S": "Start of System Hours",
+        b"Q": "Start of Market Hours",
+        b"M": "End of Market Hours",
+        b"E": "End of System Hours",
+        b"C": "End of Messages",
+    }
+
+    market = {
+        b"N": "NYSE",
+        b"A": "AMEX",
+        b"P": "Arca",
+        b"Q": "NASDAQ Global Select",
+        b"G": "NASDAQ Global Market",
+        b"S": "NASDAQ Capital Market",
+        b"Z": "BATS",
+        b" ": "Not available",
+    }
+
+    tradingStates = {
+        b"H": "Halted across all U.S. equity markets / SROs",
+        b"P": "Paused across all U.S. equity markets / SROs",
+        b"Q": "Quotation only period for cross-SRO halt or pause",
+        b"T": "Trading on NASDAQ",
+    }
+
+    primaryMarketMaker = {
+        b"Y": "Primary market maker",
+        b"N": "Non-primary market maker",
+    }
+
+    marketMakerModes = {
+        b"N": "Normal",
+        b"P": "Passive",
+        b"S": "Syndicate",
+        b"R": "Pre-syndicate",
+        b"L": "Penalty",
+    }
+
+    marketParticipantStates = {
+        b"A": "Active",
+        b"E": "Excused",
+        b"W": "Withdrawn",
+        b"S": "Suspended",
+        b"D": "Deleted",
+    }
+
+    def __init__(self) -> None:
+        """Initialize an ITCH4MarketMessage."""
+        self.timestamp: int = 0
+
+    @classmethod
+    def from_bytes(cls, message_data: bytes) -> "ITCH4MarketMessage":
+        """Create a message object from bytes data.
+
+        Args:
+            message_data: The raw message bytes
+
+        Returns:
+            The appropriate message object based on message type
+        """
+        from ..message_reader import UnknownMessageTypeError
+
+        if not message_data:
+            raise ValueError("Empty message data")
+
+        message_type = bytes([message_data[0]])
+
+        # Message type mapping for ITCH 4.0
+        message_classes = {
+            b"T": SecondsMessage,
+            b"S": SystemEventMessage,
+            b"R": StockDirectoryMessage,
+            b"H": StockTradingActionMessage,
+            b"L": MarketParticipantPositionMessage,
[…]
