+++ src/meatpy/itch3/itch3_market_processor.py
@@ -1,0 +1,426 @@
+"""ITCH 3.0 market processor for limit order book reconstruction.
+
+This module provides the ITCH3MarketProcessor class, which processes ITCH 3.0
+market messages to reconstruct the limit order book and handle trading status updates.
+"""
+
+import datetime
+
+from ..lob import LimitOrderBook, OrderNotFoundError, OrderType
+from ..market_processor import MarketProcessor
+from ..message_reader import MarketMessage
+from ..timestamp import Timestamp
+from ..trading_status import (
+    HaltedTradingStatus,
+    PostTradeTradingStatus,
+    PreTradeTradingStatus,
+    QuoteOnlyTradingStatus,
+    TradeTradingStatus,
+)
+from .itch3_market_message import (
+    AddOrderMessage,
+    AddOrderMPIDMessage,
+    BrokenTradeMessage,
+    CrossTradeMessage,
+    ITCH3MarketMessage,
+    MillisecondsMessage,
+    NoiiMessage,
+    OrderCancelMessage,
+    OrderDeleteMessage,
+    OrderExecutedMessage,
+    OrderExecutedPriceMessage,
+    SecondsMessage,
+    StockDirectoryMessage,
+    StockTradingActionMessage,
+    SystemEventMessage,
+    TradeMessage,
+)
+
+
+class InvalidBuySellIndicatorError(Exception):
+    """Exception raised when an invalid buy/sell indicator is encountered.
+
+    This exception is raised when the buy/sell indicator value is not
+    recognized by the ITCH 3.0 processor.
+    """
+
+    pass
+
+
+class MissingLOBError(Exception):
+    """Exception raised when attempting to perform operations without a LOB.
+
+    This exception is raised when trying to cancel, execute, or modify
+    orders when no limit order book is available.
+    """
+
+    pass
+
+
+class UnknownSystemMessageError(Exception):
+    """Exception raised when an unknown system message is encountered.
+
+    This exception is raised when the system message code is not
+    recognized by the ITCH 3.0 processor.
+    """
+
+    pass
+
+
+class ITCH3MarketProcessor(MarketProcessor[int, int, int, int, dict[str, str]]):
+    """A market processor for ITCH 3.0 format.
+
+    This processor handles ITCH 3.0 market messages and reconstructs the limit
+    order book. It processes order additions, executions, cancellations, and
+    trading status updates.
+
+    ITCH 3.0 is ASCII format with separate T (seconds) and M (milliseconds)
+    timestamp messages.
+
+    Attributes:
+        system_status: Current system status code
+        stock_status: Current stock trading status code
+        current_second: Current second from T message
+        current_millisecond: Current millisecond from M message
+    """
+
+    def __init__(self, instrument: str | bytes, book_date: datetime.datetime) -> None:
+        """Initialize the ITCH3MarketProcessor.
+
+        Args:
+            instrument: The instrument symbol to process
+            book_date: The date for the limit order book
+        """
+        super().__init__(instrument, book_date)
+        self.current_second: int = 0
+        self.current_millisecond: int = 0
+        self.system_status: bytes = b""
+        self.stock_status: bytes = b""
+        self.lob: LimitOrderBook[int, int, int, int, dict[str, str]] | None = None
+
+    def adjust_timestamp(self) -> Timestamp:
+        """Adjust the raw timestamp to a datetime object for ITCH 3.0.
+
+        In ITCH 3.0, timestamps are assembled from:
+        - current_second: from the latest SecondsMessage (T)
+        - current_millisecond: from the latest MillisecondsMessage (M)
+
+        Returns:
+            A Timestamp object representing the adjusted time
+        """
+        # Convert seconds and milliseconds to nanoseconds
[…]
