+++ src/meatpy/itch3/__init__.py
@@ -1,0 +1,71 @@
+"""ITCH 3.0 market data subpackage.
+
+This package provides message types, parsers, processors, and recorders for handling
+ITCH 3.0 market data in MeatPy.
+
+ITCH 3.0 is an ASCII format with separate timestamp messages:
+- T: Seconds from midnight
+- M: Milliseconds offset within current second
+
+Message Types:
+- T: Seconds
+- M: Milliseconds
+- S: System Event
+- R: Stock Directory
+- H: Stock Trading Action
+- L: Market Participant Position
+- A: Add Order (no MPID)
+- F: Add Order (with MPID)
+- E: Order Executed
+- C: Order Executed With Price
+- X: Order Cancel
+- D: Order Delete
+- P: Trade
+- Q: Cross Trade
+- B: Broken Trade
+- I: NOII
+"""
+
+from .itch3_market_message import (
+    ITCH3MarketMessage,
+    AddOrderMessage,
+    AddOrderMPIDMessage,
+    BrokenTradeMessage,
+    CrossTradeMessage,
+    MarketParticipantPositionMessage,
+    MillisecondsMessage,
+    NoiiMessage,
+    OrderCancelMessage,
+    OrderDeleteMessage,
+    OrderExecutedMessage,
+    OrderExecutedPriceMessage,
+    SecondsMessage,
+    StockDirectoryMessage,
+    StockTradingActionMessage,
+    SystemEventMessage,
+    TradeMessage,
+)
+from .itch3_market_processor import ITCH3MarketProcessor
+from .itch3_message_reader import ITCH3MessageReader
+
+__all__ = [
+    "ITCH3MarketMessage",
+    "ITCH3MarketProcessor",
+    "ITCH3MessageReader",
+    "AddOrderMessage",
+    "AddOrderMPIDMessage",
+    "BrokenTradeMessage",
+    "CrossTradeMessage",
+    "MarketParticipantPositionMessage",
+    "MillisecondsMessage",
+    "NoiiMessage",
+    "OrderCancelMessage",
+    "OrderDeleteMessage",
+    "OrderExecutedMessage",
+    "OrderExecutedPriceMessage",
+    "SecondsMessage",
+    "StockDirectoryMessage",
+    "StockTradingActionMessage",
+    "SystemEventMessage",
+    "TradeMessage",
+]

[…]
