### 2026-01-24 9fe8342
feat: add support for IEX DEEP 1.0 format
Implements GitHub issue #82: Add support for IEX DEEP format.

This adds a new `iex_deep` module with:

- Message types: All 12 IEX DEEP message types including SystemEvent,
  SecurityDirectory, TradingStatus, PriceLevelUpdate (buy/sell),
  TradeReport, OfficialPrice, TradeBreak, and AuctionInformation

- IEXDEEPMessageReader: Reads PCAP and PCAP-NG format files with
  IEX-TP transport protocol parsing. Supports compressed files.

- IEXDEEPMarketProcessor: Reconstructs order book from price-level
  updates (aggregated quotes, not individual orders like
[…]
