+++ tests/test_iex_deep_processor.py
@@ -1,0 +1,211 @@
+"""Tests for IEX DEEP market processor functionality."""
+
+import struct
+
+
+from meatpy.iex_deep.iex_deep_market_processor import IEXDEEPMarketProcessor
+from meatpy.iex_deep.iex_deep_market_message import (
+    IEXDEEPMarketMessage,
+)
+
+
+def make_buy_price_level(
+    symbol: bytes, price: int, size: int, timestamp: int = 1527599700000000000
+):
+    """Helper to create a buy price level update message."""
+    data = struct.pack("<c B q 8s I q", b"8", 0x01, timestamp, symbol, size, price)
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_sell_price_level(
+    symbol: bytes, price: int, size: int, timestamp: int = 1527599700000000000
+):
+    """Helper to create a sell price level update message."""
+    data = struct.pack("<c B q 8s I q", b"5", 0x01, timestamp, symbol, size, price)
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_trade(
+    symbol: bytes,
+    price: int,
+    size: int,
+    trade_id: int,
+    timestamp: int = 1527599700000000000,
+):
+    """Helper to create a trade report message."""
+    data = struct.pack(
+        "<c B q 8s I q q", b"T", 0x00, timestamp, symbol, size, price, trade_id
+    )
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_system_event(system_event: bytes, timestamp: int = 1527599700000000000):
+    """Helper to create a system event message."""
+    data = struct.pack("<c c q", b"S", system_event, timestamp)
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_trading_status(
+    symbol: bytes, status: bytes, timestamp: int = 1527599700000000000
+):
+    """Helper to create a trading status message."""
+    data = struct.pack("<c c q 8s 4s", b"H", status, timestamp, symbol, b"    ")
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+class TestIEXDEEPMarketProcessor:
+    """Test the IEXDEEPMarketProcessor class."""
+
+    def test_initialization(self):
+        """Test processor initialization."""
+        processor = IEXDEEPMarketProcessor("AAPL")
+        assert processor.instrument == b"AAPL    "
+        assert processor._bid_levels == {}
+        assert processor._ask_levels == {}
+        assert processor.track_lob is False
+
+    def test_initialization_with_bytes_symbol(self):
+        """Test processor initialization with bytes symbol."""
+        processor = IEXDEEPMarketProcessor(b"SPY     ")
+        assert processor.instrument == b"SPY     "
+
+    def test_price_level_update_bid(self):
+        """Test processing bid price level update."""
+        processor = IEXDEEPMarketProcessor("AAPL")
+
+        message = make_buy_price_level(b"AAPL    ", 1850000, 1000)
+        processor.process_message(message)
+
+        assert 1850000 in processor._bid_levels
+        assert processor._bid_levels[1850000] == 1000
+
+    def test_price_level_update_ask(self):
+        """Test processing ask price level update."""
+        processor = IEXDEEPMarketProcessor("AAPL")
+
+        message = make_sell_price_level(b"AAPL    ", 1851000, 500)
+        processor.process_message(message)
+
+        assert 1851000 in processor._ask_levels
+        assert processor._ask_levels[1851000] == 500
+
+    def test_price_level_removal(self):
+        """Test that size=0 removes a price level."""
+        processor = IEXDEEPMarketProcessor("AAPL")
+
+        # Add a price level
+        message1 = make_buy_price_level(b"AAPL    ", 1850000, 1000)
+        processor.process_message(message1)
+
[…]
