+++ src/meatpy/iex_deep/iex_deep_market_processor.py
@@ -1,0 +1,474 @@
+"""IEX DEEP market processor for price level-based order book reconstruction.
+
+This module provides the IEXDEEPMarketProcessor class, which processes IEX DEEP
+market messages to reconstruct the limit order book based on price level updates.
+
+Unlike ITCH 5.0 which provides individual orders, IEX DEEP provides aggregated
+price level updates, so the processor works with price levels directly.
+"""
+
+from __future__ import annotations
+
+import datetime
+from typing import Optional
+
+from ..lob import LimitOrderBook, OrderType
+from ..market_processor import MarketProcessor
+from ..message_reader import MarketMessage
+from ..timestamp import Timestamp
+from ..trading_status import (
+    HaltedTradingStatus,
+    PostTradeTradingStatus,
+    PreTradeTradingStatus,
+    TradeTradingStatus,
+)
+from ..types import Price, TradeRef, Volume
+from .iex_deep_market_message import (
+    AuctionInformationMessage,
+    IEXDEEPMarketMessage,
+    OperationalHaltStatusMessage,
+    OfficialPriceMessage,
+    PriceLevelUpdateBuySideMessage,
+    PriceLevelUpdateSellSideMessage,
+    SecurityDirectoryMessage,
+    SecurityEventMessage,
+    ShortSalePriceTestStatusMessage,
+    SystemEventMessage,
+    TradeBreakMessage,
+    TradeReportMessage,
+    TradingStatusMessage,
+)
+
+
+class InvalidMessageTypeError(Exception):
+    """Exception raised when an invalid message type is encountered."""
+
+    pass
+
+
+class IEXDEEPMarketProcessor(MarketProcessor[int, int, int, int, dict[str, str]]):
+    """A market processor for IEX DEEP format.
+
+    This processor handles IEX DEEP market messages and reconstructs the limit
+    order book based on price level updates. Unlike ITCH 5.0, IEX DEEP provides
+    aggregated price levels directly rather than individual orders.
+
+    Attributes:
+        system_status: Current system event status
+        trading_status_code: Current trading status code for the instrument
+        operational_halt_status: Current operational halt status
+        short_sale_status: Current short sale price test status
+    """
+
+    def __init__(
+        self,
+        instrument: str | bytes,
+        book_date: Optional[datetime.datetime] = None,
+        track_lob: bool = False,
+    ) -> None:
+        """Initialize the IEXDEEPMarketProcessor.
+
+        Args:
+            instrument: The instrument/symbol to process (8 chars, space-padded)
+            book_date: The trading date for this processor (optional for IEX DEEP
+                       since timestamps are absolute)
+            track_lob: Whether to track the full LOB (default False for IEX DEEP
+                       since it uses price-level data, not individual orders)
+        """
+        # book_date is optional for IEX DEEP since timestamps are POSIX epoch
+        if book_date is None:
+            book_date = datetime.datetime.now(datetime.timezone.utc)
+
+        super(IEXDEEPMarketProcessor, self).__init__(instrument, book_date)
+        # IEX DEEP doesn't use order-level LOB tracking by default
+        self.track_lob = track_lob
+
+        # Override instrument to store as bytes with proper formatting
+        self.instrument: bytes = (
+            f"{instrument:<8}".encode("ascii")
+            if isinstance(instrument, str)
+            else instrument
+        )
+        self.book_date: datetime.datetime = book_date
+        self.system_status: bytes = b""
+        self.trading_status_code: bytes = b""
+        self.operational_halt_status: bytes = b""
+        self.short_sale_status: int = 0
+
+        self._trade_ids: set[int] = set()
+        # Track price levels: {price: size}
[…]
