### 2026-01-24 9c8ddc6
test: increase IEX DEEP test coverage from 46% to 72%
Added comprehensive tests for:
- All 12 IEX DEEP message types (from_bytes, to_bytes, to_json)
- Message property methods (flag checks, sale condition flags)
- Processor trading status transitions (halted, paused, trading, etc.)
- System event processing (start/end of messages, hours)
- Operational halt processing
- Short sale status processing
- Trade break processing
- Security directory, auction info, official price messages
- Symbol filtering for all message types

Coverage breakdown:
- iex_deep_market_message.py: 99.30%
- iex_deep_mark
[…]
