+++ tests/test_iex_deep_processor.py
@@ -2,10 +2,16 @@
 
 import struct
 
-
-from meatpy.iex_deep.iex_deep_market_processor import IEXDEEPMarketProcessor
-from meatpy.iex_deep.iex_deep_market_message import (
-    IEXDEEPMarketMessage,
+from meatpy.iex_deep.iex_deep_market_processor import (
+    IEXDEEPMarketProcessor,
+    InvalidMessageTypeError,
+)
+from meatpy.iex_deep.iex_deep_market_message import IEXDEEPMarketMessage
+from meatpy.trading_status import (
+    HaltedTradingStatus,
+    PreTradeTradingStatus,
+    TradeTradingStatus,
+    PostTradeTradingStatus,
 )
 
 
@@ -39,6 +45,20 @@ def make_trade(
     return IEXDEEPMarketMessage.from_bytes(data)
 
 
+def make_trade_break(
+    symbol: bytes,
+    price: int,
+    size: int,
+    trade_id: int,
+    timestamp: int = 1527599700000000000,
+):
+    """Helper to create a trade break message."""
+    data = struct.pack(
+        "<c B q 8s I q q", b"B", 0x00, timestamp, symbol, size, price, trade_id
+    )
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
 def make_system_event(system_event: bytes, timestamp: int = 1527599700000000000):
     """Helper to create a system event message."""
     data = struct.pack("<c c q", b"S", system_event, timestamp)
@@ -50,6 +70,78 @@ def make_trading_status(
 ):
     """Helper to create a trading status message."""
     data = struct.pack("<c c q 8s 4s", b"H", status, timestamp, symbol, b"    ")
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_operational_halt(
+    symbol: bytes, status: bytes, timestamp: int = 1527599700000000000
+):
+    """Helper to create an operational halt message."""
+    data = struct.pack("<c c q 8s", b"O", status, timestamp, symbol)
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_short_sale_status(
+    symbol: bytes, status: int, timestamp: int = 1527599700000000000
+):
+    """Helper to create a short sale status message."""
+    data = struct.pack("<c B q 8s c", b"P", status, timestamp, symbol, b"A")
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_security_event(
+    symbol: bytes, event: bytes, timestamp: int = 1527599700000000000
+):
+    """Helper to create a security event message."""
+    data = struct.pack("<c c q 8s", b"E", event, timestamp, symbol)
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_official_price(
+    symbol: bytes, price_type: bytes, price: int, timestamp: int = 1527599700000000000
+):
+    """Helper to create an official price message."""
+    data = struct.pack("<c c q 8s q", b"X", price_type, timestamp, symbol, price)
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_security_directory(symbol: bytes, timestamp: int = 1527599700000000000):
+    """Helper to create a security directory message."""
+    data = struct.pack(
+        "<c B q 8s I q B",
+        b"D",
+        0x00,  # flags
+        timestamp,
+        symbol,
+        100,  # round_lot_size
+        1500000,  # adjusted_poc_price
+        1,  # luld_tier
+    )
+    return IEXDEEPMarketMessage.from_bytes(data)
+
+
+def make_auction_info(
+    symbol: bytes, auction_type: bytes, timestamp: int = 1527599700000000000
+):
+    """Helper to create an auction information message."""
[…]
