+++ tests/test_iex_deep_messages.py
@@ -10,10 +10,17 @@ from meatpy.iex_deep.iex_deep_market_message import (
     SystemEventMessage,
     SecurityDirectoryMessage,
     TradingStatusMessage,
+    OperationalHaltStatusMessage,
+    ShortSalePriceTestStatusMessage,
+    SecurityEventMessage,
     PriceLevelUpdateBuySideMessage,
     PriceLevelUpdateSellSideMessage,
     TradeReportMessage,
+    OfficialPriceMessage,
+    TradeBreakMessage,
+    AuctionInformationMessage,
 )
+from meatpy.message_reader import UnknownMessageTypeError
 
 
 class TestSystemEventMessage:
@@ -21,14 +28,10 @@ class TestSystemEventMessage:
 
     def test_from_bytes(self):
         """Test creating message from bytes."""
-        # IEX DEEP format: message_type(1) + system_event(1) + timestamp(8)
-        timestamp = 1527599700000000000  # nanoseconds since POSIX epoch
-        system_event = b"O"  # Start of messages
+        timestamp = 1527599700000000000
+        system_event = b"O"
 
-        # Pack the data in little endian
         data = struct.pack("<c c q", b"S", system_event, timestamp)
-
-        # Create message from bytes
         message = IEXDEEPMarketMessage.from_bytes(data)
 
         assert isinstance(message, SystemEventMessage)
@@ -38,17 +41,13 @@ class TestSystemEventMessage:
     def test_to_bytes(self):
         """Test converting message to bytes."""
         message = SystemEventMessage(system_event=b"O", timestamp=1527599700000000000)
-
         data = message.to_bytes()
-
-        # Verify the packed data
         expected = struct.pack("<c b q", b"S", ord(b"O"), 1527599700000000000)
         assert data == expected
 
     def test_to_json(self):
         """Test JSON serialization."""
         message = SystemEventMessage(system_event=b"O", timestamp=1527599700000000000)
-
         json_str = message.to_json()
         data = json.loads(json_str)
 
@@ -56,19 +55,27 @@ class TestSystemEventMessage:
         assert data["message_type"] == "S"
         assert data["system_event"] == "O"
 
+    def test_roundtrip(self):
+        """Test that from_bytes(to_bytes()) returns equivalent message."""
+        original = SystemEventMessage(system_event=b"R", timestamp=1527599700000000000)
+        data = original.to_bytes()
+        restored = IEXDEEPMarketMessage.from_bytes(data)
+
+        assert isinstance(restored, SystemEventMessage)
+        assert restored.timestamp == original.timestamp
+        assert restored.system_event == original.system_event
+
 
 class TestSecurityDirectoryMessage:
     """Test SecurityDirectoryMessage functionality."""
 
     def test_from_bytes(self):
         """Test creating message from bytes."""
-        # Format: message_type(1) + flags(1) + timestamp(8) + symbol(8) +
-        #         round_lot_size(4) + adjusted_poc_price(8) + luld_tier(1)
         timestamp = 1527599700000000000
-        flags = 0x80  # Test flag bit
+        flags = 0x80
         symbol = b"AAPL    "
         round_lot_size = 100
-        adjusted_poc_price = 1500000  # $150.0000 (4 decimal places)
+        adjusted_poc_price = 1500000
         luld_tier = 1
 
         data = struct.pack(
@@ -92,19 +99,282 @@ class TestSecurityDirectoryMessage:
         assert message.adjusted_poc_price == adjusted_poc_price
         assert message.luld_tier == luld_tier
 
+    def test_to_bytes(self):
+        """Test converting message to bytes."""
+        message = SecurityDirectoryMessage(
+            flags=0x80,
[…]
