### 2026-01-24 9820414
test: add comprehensive tests for ITCH 2.0, 3.0, and 4.0 modules
- Add 178 tests covering message parsing, file reading, and LOB reconstruction
- Achieve 80.51% code coverage for new ITCH modules
- Fix bug in Timestamp.from_datetime() missing second field
- Apply ruff formatting to source files

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
