+++ tests/test_itch2_messages.py
@@ -1,0 +1,348 @@
+"""Tests for ITCH 2.0 market message functionality."""
+
+import json
+import pytest
+
+from meatpy.itch2.itch2_market_message import (
+    ITCH2MarketMessage,
+    SystemEventMessage,
+    AddOrderMessage,
+    OrderExecutedMessage,
+    OrderCancelMessage,
+    TradeMessage,
+    BrokenTradeMessage,
+)
+
+
+class TestITCH2MarketMessage:
+    """Test the base ITCH2MarketMessage class."""
+
+    def test_from_bytes_unknown_type(self):
+        """Test handling of unknown message types."""
+        data = b"12345678Z" + b"0" * 20
+        with pytest.raises(ValueError, match="Unknown message type"):
+            ITCH2MarketMessage.from_bytes(data)
+
+    def test_from_bytes_too_short(self):
+        """Test handling of message that is too short."""
+        data = b"1234"
+        with pytest.raises(ValueError, match="Message too short"):
+            ITCH2MarketMessage.from_bytes(data)
+
+    def test_from_json_unknown_type(self):
+        """Test handling of unknown message type in JSON."""
+        json_str = json.dumps({"message_type": "Z", "timestamp": 12345})
+        with pytest.raises(ValueError, match="Unknown message type"):
+            ITCH2MarketMessage.from_json(json_str)
+
+
+class TestSystemEventMessage:
+    """Test SystemEventMessage functionality."""
+
+    def test_from_bytes_data(self):
+        """Test creating message from bytes."""
+        # Format: timestamp(8) + type(1) + event_code(1) = 10 chars
+        data = b"12345678SO"  # timestamp=12345678, type=S, event_code=O
+
+        message = ITCH2MarketMessage.from_bytes(data)
+
+        assert isinstance(message, SystemEventMessage)
+        assert message.timestamp == 12345678
+        assert message.event_code == b"O"
+
+    def test_to_bytes(self):
+        """Test converting message to bytes."""
+        message = SystemEventMessage()
+        message.timestamp = 12345678
+        message.event_code = b"O"
+
+        data = message.to_bytes()
+
+        assert data == b"12345678SO"
+
+    def test_json_roundtrip(self):
+        """Test JSON serialization and deserialization."""
+        message = SystemEventMessage()
+        message.timestamp = 12345678
+        message.event_code = b"Q"
+
+        json_str = message.to_json()
+        data = json.loads(json_str)
+
+        assert data["timestamp"] == 12345678
+        assert data["message_type"] == "S"
+        assert data["event_code"] == "Q"
+
+        # Roundtrip
+        new_message = ITCH2MarketMessage.from_json(json_str)
+        assert isinstance(new_message, SystemEventMessage)
+        assert new_message.timestamp == message.timestamp
+        assert new_message.event_code == message.event_code
+
+    def test_all_event_codes(self):
+        """Test all valid event codes."""
+        event_codes = [b"O", b"S", b"Q", b"M", b"E", b"C"]
+        for code in event_codes:
+            data = f"12345678S{code.decode()}".encode()
+            message = ITCH2MarketMessage.from_bytes(data)
+            assert message.event_code == code
+
+
+class TestAddOrderMessage:
+    """Test AddOrderMessage functionality."""
+
+    def test_from_bytes_data(self):
+        """Test creating message from bytes."""
+        # Format: timestamp(8) + type(1) + order_ref(9) + side(1) + shares(6)
+        #         + stock(6) + price(10) + display(1) = 42 chars
+        data = b"12345678A000000001B000100AAPL  0000150000Y"
+
+        message = ITCH2MarketMessage.from_bytes(data)
+
+        assert isinstance(message, AddOrderMessage)
[…]
