+++ tests/test_itch3_messages.py
@@ -1,0 +1,537 @@
+"""Tests for ITCH 3.0 market message functionality."""
+
+import json
+import pytest
+
+from meatpy.itch3.itch3_market_message import (
+    ITCH3MarketMessage,
+    SecondsMessage,
+    MillisecondsMessage,
+    SystemEventMessage,
+    StockDirectoryMessage,
+    StockTradingActionMessage,
+    MarketParticipantPositionMessage,
+    AddOrderMessage,
+    AddOrderMPIDMessage,
+    OrderExecutedMessage,
+    OrderExecutedPriceMessage,
+    OrderCancelMessage,
+    OrderDeleteMessage,
+    TradeMessage,
+    CrossTradeMessage,
+    BrokenTradeMessage,
+    NoiiMessage,
+)
+
+
+class TestITCH3MarketMessage:
+    """Test the base ITCH3MarketMessage class."""
+
+    def test_from_bytes_unknown_type(self):
+        """Test handling of unknown message types."""
+        data = b"Z" + b"0" * 20
+        with pytest.raises(ValueError, match="Unknown message type"):
+            ITCH3MarketMessage.from_bytes(data)
+
+    def test_from_bytes_too_short(self):
+        """Test handling of message that is too short."""
+        data = b""
+        with pytest.raises(ValueError, match="Message too short"):
+            ITCH3MarketMessage.from_bytes(data)
+
+    def test_from_json_unknown_type(self):
+        """Test handling of unknown message type in JSON."""
+        json_str = json.dumps({"message_type": "Z"})
+        with pytest.raises(ValueError, match="Unknown message type"):
+            ITCH3MarketMessage.from_json(json_str)
+
+
+class TestSecondsMessage:
+    """Test SecondsMessage functionality."""
+
+    def test_from_bytes_data(self):
+        """Test creating message from bytes."""
+        # Format: T + seconds(5) = 6 chars
+        data = b"T12345"
+
+        message = ITCH3MarketMessage.from_bytes(data)
+
+        assert isinstance(message, SecondsMessage)
+        assert message.seconds == 12345
+
+    def test_to_bytes(self):
+        """Test converting message to bytes."""
+        message = SecondsMessage()
+        message.seconds = 12345
+
+        data = message.to_bytes()
+
+        assert data == b"T12345"
+
+    def test_json_roundtrip(self):
+        """Test JSON serialization and deserialization."""
+        message = SecondsMessage()
+        message.seconds = 54321
+
+        json_str = message.to_json()
+        new_message = ITCH3MarketMessage.from_json(json_str)
+
+        assert isinstance(new_message, SecondsMessage)
+        assert new_message.seconds == message.seconds
+
+
+class TestMillisecondsMessage:
+    """Test MillisecondsMessage functionality."""
+
+    def test_from_bytes_data(self):
+        """Test creating message from bytes."""
+        # Format: M + milliseconds(3) = 4 chars
+        data = b"M123"
+
+        message = ITCH3MarketMessage.from_bytes(data)
+
+        assert isinstance(message, MillisecondsMessage)
+        assert message.milliseconds == 123
+
+    def test_to_bytes(self):
+        """Test converting message to bytes."""
+        message = MillisecondsMessage()
+        message.milliseconds = 456
+
+        data = message.to_bytes()
+
+        assert data == b"M456"
+
+    def test_json_roundtrip(self):
+        """Test JSON serialization and deserialization."""
+        message = MillisecondsMessage()
+        message.milliseconds = 789
+
+        json_str = message.to_json()
+        new_message = ITCH3MarketMessage.from_json(json_str)
+
+        assert isinstance(new_message, MillisecondsMessage)
+        assert new_message.milliseconds == message.milliseconds
+
+
+class TestSystemEventMessage:
[…]
