### 2026-01-24 255aedc
chore: bump version to 0.3.0
