### 2026-01-24 cb573e4
chore: bump version to 0.4.0
