### 2026-03-05 8b8c5d2
Updated "State of the Field" section
