### 2026-04-28 #10480 @editorialbot
[REVIEW]: MeatPy: A Python Framework for Limit Order Book Reconstruction and Analysis
**Submitting author:** @vgreg (<a href="http://orcid.org/0000-0002-1230-5026">Vincent Grégoire</a>)
**Repository:** https://github.com/vgreg/MeatPy
**Branch with paper.md** (empty if default branch): 64-submit-paper-to-joss
**Version:** v0.2.7
