### 2026-05-02 #10480 @robertmartin8
- I was able to install and run all tests.
