### 2026-05-02 #10480 @robertmartin8
I have completed my review. This was a fast review because the paper is well-written, there is a real scholarly use case, and this presents as a mature  library, built with good OSS practices, that has clearly supported academic research.

I thought the Software Design section thoughtfully explained the real tradeoffs. I initially felt that the "Challenges in Processing Limit Order Book Data" section was too long, but my present judgment is that it provides necessary context for the rest of the paper and I recommend it be kept largely as is.

Here some small improvement points @vgreg:

- The docs' verification command import meatpy; print(meatpy.__version__) fails because meatpy.__version__ is not defined. There is also some version metadata inconsistency: review issue says v0.2.7, pyproject.toml says 0.2.10, uv.lock says 0.2.8, and latest PyPI is 0.4.0.
- I would recommend including a small sample of data (possibly faked if licensing is an issue).
- A few small typos in the paper:
    - "developement" -> "development"
    - "such a generators" -> "such as generators"
    - "during from the time they became available" needs rewriting
    - "Subsequent messages referencing the same order—such as cancellations or executions omit..."  needs rewriting
    - "PyPi Stats" -> "PyPI Stats"
