### 2026-05-06 eb2095a
Address JOSS reviewer comments on paper
Fixed typos and updated the supported-formats footnote to reflect
current ITCH (2.0-5.0) and IEX DEEP support.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
