### 2026-05-06 03500cb
docs: ship synthetic ITCH 5.0 sample and quickstart
Adds samples/data/sample.itch50 (590 bytes, 20 messages, fictional
symbols MEAT and PYTH) so users can run the parser, processor, and
LOB end-to-end without a Nasdaq subscription. Includes the generator
script, a README, and samples/itch50/00_quickstart.py demonstrating
message-count and top-of-book reconstruction against the sample.

Addresses the JOSS reviewer's request for bundled sample data.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
