### 2026-05-12 f325393
docs: surface bundled ITCH 5.0 quickstart and fix LOB example
Add a top-level pointer in the getting-started guide and the landing page to
the synthetic samples/data/sample.itch50 + samples/itch50/00_quickstart.py
end-to-end flow so users can try MeatPy without a Nasdaq subscription. Fix
the LOB code snippet to use `processor.current_lob` (the actual attribute)
and drop the redundant `/ 10000` since `best_bid`/`best_ask` already return
a Decimal scaled to dollars.

Co-Authored-By: Claude Opus 4.7 (1M context) <noreply@anthropic.com>
