### 2026-05-12 3fffbf6
chore: bump version to 0.5.0
