### 2026-06-05 c533a98
chore(deps): bump pyarrow from 20.0.0 to 23.0.1
Bumps [pyarrow](https://github.com/apache/arrow) from 20.0.0 to 23.0.1.
- [Release notes](https://github.com/apache/arrow/releases)
- [Commits](https://github.com/apache/arrow/compare/apache-arrow-20.0.0...apache-arrow-23.0.1)

---
updated-dependencies:
- dependency-name: pyarrow
  dependency-version: 23.0.1
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
