+++ pyproject.toml
@@ -10,7 +10,7 @@ maintainers = [{ name = "Vincent Grégoire", email = "vincent.gregoire@hec.ca" }
 readme = "README.md"
 license = { text = "BSD-3-Clause" }
 requires-python = ">=3.11"
-dependencies = ["pyarrow>=18.0.0"]
+dependencies = ["pyarrow>=10.0.0"]
 classifiers = [
   "Intended Audience :: Education",
   "Intended Audience :: Financial and Insurance Industry",
